requirements.txt
imgaug
opencv-python
lxml
torch
torchvision
torchsummary
wandb
Pillow
tqdm
yaml
